app.py
from fastapi import FastAPI, File, UploadFile
from fastapi.middleware.cors import CORSMiddleware
import tensorflow as tf
from tensorflow import keras
from tensorflow.keras.applications.mobilenet_v2 import preprocess_input
import numpy as np
from PIL import Image
import io
import uvicorn
import warnings
import traceback

warnings.filterwarnings('ignore')

app = FastAPI()

# Allow CORS for mobile app
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# ============================================
# DEFINE CUSTOM LOSS FUNCTION FOR HYBRID MODEL ✅
# ============================================
def label_smoothing_loss(y_true, y_pred, smoothing=0.1):
    """Label smoothing for better generalization"""
    num_classes = tf.cast(tf.shape(y_true)[-1], tf.float32)
    y_true_smooth = y_true * (1 - smoothing) + (smoothing / num_classes)
    return keras.losses.categorical_crossentropy(y_true_smooth, y_pred)

# ============================================
# LOAD YOUR 3 MODELS (.h5 FORMAT) ✅
# ============================================
print("Loading models...")

try:
    model_mobilenet = keras.models.load_model('models/final_mobilenetv2_model.h5')
    print("✅ MobileNetV2 loaded")
except Exception as e:
    print(f"❌ Error loading MobileNetV2: {str(e)[:100]}")
    model_mobilenet = None

try:
    model_densenet = keras.models.load_model('models/final_densenet121_model.h5')
    print("✅ DenseNet121 loaded")
except Exception as e:
    print(f"❌ Error loading DenseNet121: {str(e)[:100]}")
    model_densenet = None

try:
    # ✅ FIX: Load with custom loss function
    model_hybrid = keras.models.load_model(
        'models/best_enhanced_hybrid_finetuned.h5',
        custom_objects={'label_smoothing_loss': label_smoothing_loss}
    )
    print("✅ Hybrid Model loaded")
except Exception as e:
    print(f"❌ Error loading Hybrid Model: {str(e)[:100]}")
    model_hybrid = None

# ============================================
# CLASS LABELS (MATCH TRAINING ORDER!) ✅
# ============================================
CLASS_LABELS = [
    'Acne',
    'Eczema',
    'Keratosis',
    'Carcinoma',
    'Milia',
    'Rosacea',
]

# ============================================
# IMAGE PREPROCESSING ✅ MobileNetV2 specific
# ============================================
def preprocess_image(image_bytes):
    """Convert image bytes to preprocessed tensor"""
    try:
        print(f"[PREPROCESS] Received {len(image_bytes)} bytes")
        
        img = Image.open(io.BytesIO(image_bytes)).convert('RGB')
        print(f"[PREPROCESS] Image opened: {img.format} {img.size}")
        
        img = img.resize((224, 224))
        print(f"[PREPROCESS] Image resized to 224x224")
        
        img_array = np.array(img, dtype=np.float32)
        print(f"[PREPROCESS] Array shape: {img_array.shape}")
        
        # ✅ Use MobileNetV2 preprocessing ([-1, 1] range)
        img_array = preprocess_input(img_array)
        print(f"[PREPROCESS] Preprocessing applied, range: [{img_array.min():.2f}, {img_array.max():.2f}]")
        
        img_array = np.expand_dims(img_array, axis=0)
        print(f"[PREPROCESS] Final shape: {img_array.shape}")
        
        return img_array
    except Exception as e:
        print(f"[PREPROCESS ERROR] {e}")
        traceback.print_exc()
        return None

# ============================================
# GET PREDICTIONS FROM ONE MODEL
# ============================================
def get_model_predictions(model, image_array, model_name, model_type):
    """Run inference on one model"""
    if model is None:
        print(f"[{model_name}] Model not loaded")
        return None
    
    try:
        print(f"[{model_name}] Running prediction...")
        predictions = model.predict(image_array, verbose=0)
        print(f"[{model_name}] Predictions shape: {predictions.shape}")
        
        confidences = predictions[0] * 100
        print(f"[{model_name}] Confidences: {confidences}")
        
        conditions = []
        for i, class_name in enumerate(CLASS_LABELS):
            conditions.append({
                'name': class_name,
                'confidence': float(confidences[i]),
                'description': get_condition_description(class_name),
                'recommendations': get_recommendations(class_name)
            })
        
        conditions.sort(key=lambda x: x['confidence'], reverse=True)
        
        metrics = {
            'accuracy': 0.95,
            'precision': 0.93,
            'f1Score': 0.94
        }
        
        result = {
            'modelName': model_name,
            'modelType': model_type,
            'metrics': metrics,
            'predictions': conditions
        }
        
        print(f"[{model_name}] ✅ Prediction successful")
        return result
    
    except Exception as e:
        print(f"[{model_name}] ❌ Error in prediction: {e}")
        traceback.print_exc()
        return None

# ============================================
# CONDITION DESCRIPTIONS
# ============================================
def get_condition_description(condition_name):
    descriptions = {
        'Acne': 'Inflammatory skin condition characterized by comedones, papules, pustules, or cysts.',
        'Eczema': 'Atopic dermatitis causing itchy, inflamed, and sometimes scaly patches of skin.',
        'Rosacea': 'Chronic inflammatory condition causing facial redness, visible blood vessels, and sometimes pustules.',
        'Keratosis': 'Rough, scaly patches caused by buildup of keratin, often from sun damage.',
        'Carcinoma': 'Abnormal growth that may indicate skin cancer. Requires immediate medical evaluation.',
        'Milia': 'Small, white keratin-filled cysts that appear as tiny bumps on the skin.',
    }
    return descriptions.get(condition_name, 'Skin condition detected. Consult a dermatologist.')

# ============================================
# RECOMMENDATIONS
# ============================================
def get_recommendations(condition_name):
    recommendations = {
        'Acne': [
            'Use non-comedogenic products',
            'Consider salicylic acid or benzoyl peroxide treatments',
            'Consult a dermatologist for persistent or severe acne'
        ],
        'Eczema': [
            'Keep skin well-moisturized with thick emollients',
            'Avoid harsh soaps and known allergens',
            'Use prescribed topical corticosteroids if recommended by a doctor'
        ],
        'Rosacea': [
            'Avoid triggers like spicy foods, alcohol, and extreme temperatures',
            'Use gentle, fragrance-free products',
            'Consider prescription treatments from a dermatologist'
        ],
        'Keratosis': [
            'Use daily broad-spectrum sunscreen (SPF 30+)',
            'Consider retinoid creams or chemical exfoliants',
            'See a dermatologist for evaluation and possible removal'
        ],
        'Carcinoma': [
            'URGENT: Schedule an appointment with a dermatologist immediately',
            'Do not delay seeking professional medical evaluation',
            'Avoid sun exposure and always use sunscreen'
        ],
        'Milia': [
            'Avoid picking or squeezing the bumps',
            'Use gentle exfoliants with AHA or BHA',
            'Consider professional extraction by a dermatologist'
        ],
    }
    return recommendations.get(condition_name, ['Consult a dermatologist for proper diagnosis'])

# ============================================
# HEALTH CHECK ENDPOINT
# ============================================
@app.get("/health")
async def health():
    print("[HEALTH] Health check requested")
    return {
        "status": "OK",
        "message": "Skin Analyzer API is running",
        "models_loaded": {
            "mobilenet": model_mobilenet is not None,
            "densenet": model_densenet is not None,
            "hybrid": model_hybrid is not None
        }
    }

# ============================================
# MAIN ANALYSIS ENDPOINT
# ============================================
@app.post("/analyze")
async def analyze_skin(image: UploadFile = File(...)):
    try:
        print("\n" + "="*60)
        print(f"[ANALYZE] Received image: {image.filename}")
        print(f"[ANALYZE] Content-Type: {image.content_type}")
        print(f"[ANALYZE] File size: {image.size} bytes" if image.size else "[ANALYZE] File size: unknown")
        
        image_bytes = await image.read()
        print(f"[ANALYZE] Read {len(image_bytes)} bytes from file")
        
        img_array = preprocess_image(image_bytes)
        
        if img_array is None:
            print("[ANALYZE] ❌ Preprocessing failed")
            return {"error": "Failed to process image"}
        
        results = []
        
        if model_mobilenet:
            print("[ANALYZE] Running MobileNetV2...")
            mobilenet_result = get_model_predictions(model_mobilenet, img_array, 'MobileNetV2', 'Baseline 1')
            if mobilenet_result:
                results.append(mobilenet_result)
        else:
            print("[ANALYZE] ⚠️  MobileNetV2 not loaded")
        
        if model_densenet:
            print("[ANALYZE] Running DenseNet121...")
            densenet_result = get_model_predictions(model_densenet, img_array, 'DenseNet-121', 'Baseline 2')
            if densenet_result:
                results.append(densenet_result)
        else:
            print("[ANALYZE] ⚠️  DenseNet121 not loaded")
        
        if model_hybrid:
            print("[ANALYZE] Running Hybrid Model...")
            hybrid_result = get_model_predictions(model_hybrid, img_array, 'Hybrid Model', 'Proposed Model')
            if hybrid_result:
                results.append(hybrid_result)
        else:
            print("[ANALYZE] ⚠️  Hybrid Model not loaded")
        
        if not results:
            print("[ANALYZE] ❌ All models failed to produce results")
            return {"error": "All models failed to produce results"}
        
        print(f"[ANALYZE] ✅ Analysis complete. Returning {len(results)} model results.")
        print("="*60 + "\n")
        return results
    
    except Exception as e:
        print(f"[ANALYZE] ❌ ERROR: {e}")
        traceback.print_exc()
        print("="*60 + "\n")
        return {"error": str(e)}

# ============================================
# RUN SERVER
# ============================================
if __name__ == "__main__":
    print("\n" + "="*50)
    print("🚀 Skin Analyzer Backend Starting...")
    print("="*50)
    print("API will run at: http://0.0.0.0:8000")
    print("Health check: http://localhost:8000/health")
    print("="*50 + "\n")
    
    uvicorn.run(app, host="0.0.0.0", port=8000)