requirements.txt
tensorflow==2.13.0
keras==2.13.1
numpy==1.24.3
scikit-learn==1.7.2
matplotlib==3.10.8
seaborn==0.13.2
fastapi==0.103.2
uvicorn==0.40.0
pillow==12.1.0
h5py==3.15.1
pandas==2.3.3
python-multipart==0.0.22